requirements.txt
requests>=2.25.1
beautifulsoup4>=4.9.3
PyYAML>=5.4.1